fix: change static_cast<int> to static_cast<size_t> for revRead call

+++ src/retractor/lib/streamInstance.cpp
@@ -108,7 +108,7 @@ rdb::payload streamInstance::constructAgsePayload(const int length,             
     auto readPosition{recordsCountSrc - fp.quot - 1};
     // Guard negative quotient before revRead: out-of-range windows must stay null, not stale 0.
     if (fp.quot >= 0 && static_cast<size_t>(fp.quot) < recordsCountSrc) {
-      if (!source->revRead(static_cast<int>(readPosition))) fp.rem = -1;
+      if (!source->revRead(static_cast<size_t>(readPosition))) fp.rem = -1;
     } else
       fp.rem = -1;  // skip to undefined(-1) as value
